pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = ""
authors = ["Sergunin Vitaliy <sergunin_v@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flake8 = "^7.3.0"
django = "^5.2.8"
psycopg = {extras = ["binary"], version = "^3.2.12"}
djangorestframework = "^3.16.1"
pillow = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
